# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Song library API by Ilya Valentuikevich",
        "title": "songLibraryAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/library/songs": {
            "get": {
                "description": "**Получения списка песен**",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Song"
                ],
                "summary": "Получения списка песен",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы",
                        "name": "groupName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название песни",
                        "name": "songTitle",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выпуска песни",
                        "name": "releaseDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ссылка на песню",
                        "name": "link",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фрагмент текста песни",
                        "name": "lyrics",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Страница",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ограничение вывода",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/models.SongsList"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/v1/library/songs/:id/lyrics": {
            "get": {
                "description": "**Получение текста песни**",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Song"
                ],
                "summary": "Получение текста песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Страница",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ограничение вывода",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lyrics"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/v1/library/songs/add": {
            "post": {
                "description": "**Добавить песню**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Song"
                ],
                "summary": "Добавить песню",
                "parameters": [
                    {
                        "description": "Информация о песне",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/utils.RespOK"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Песня уже существует",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/v1/library/songs/delete/:id": {
            "delete": {
                "description": "**Удаление песни**",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Song"
                ],
                "summary": "Удаление песни",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/utils.RespOK"
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/v1/library/songs/edit/:id": {
            "put": {
                "description": "**Редактирование данных песни, текста**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Song"
                ],
                "summary": "Редактирование песни",
                "parameters": [
                    {
                        "description": "Новая информация о песне",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Edit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/models.Song"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "песня не найдена",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Edit": {
            "type": "object",
            "properties": {
                "group_name": {
                    "type": "string",
                    "example": "Fall Out Boys"
                },
                "link": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=LBr7kECsjcQ"
                },
                "lyrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Lyrics"
                    }
                },
                "release_date": {
                    "type": "string",
                    "example": "01.01.2019"
                },
                "title": {
                    "type": "string",
                    "example": "Centuries"
                }
            }
        },
        "models.Input": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Fall Out Boys"
                },
                "song": {
                    "type": "string",
                    "example": "Centuries"
                }
            }
        },
        "models.Lyrics": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-11-23 18:55:28.896205+03"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "song_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-11-23 18:55:28.896205+03"
                },
                "verse": {
                    "type": "string",
                    "example": "Some legends are told"
                }
            }
        },
        "models.Song": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-11-23 18:55:28.896205+03"
                },
                "group_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "link": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=LBr7kECsjcQ"
                },
                "lyrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Lyrics"
                    }
                },
                "release_date": {
                    "type": "string",
                    "example": "01.01.2019"
                },
                "title": {
                    "type": "string",
                    "example": "Centuries"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-11-23 18:55:28.896205+03"
                }
            }
        },
        "models.SongsList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Song"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_count": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "utils.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "описание ошибки"
                },
                "method": {
                    "type": "string",
                    "example": "POST"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "time": {
                    "type": "string",
                    "example": "2024-11-24 19:33:57"
                },
                "title": {
                    "type": "string",
                    "example": "DB error"
                },
                "type": {
                    "type": "string",
                    "example": "/api/v1/..."
                }
            }
        },
        "utils.RespOK": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Запрос успешно выполнен"
                }
            }
        }
    }
}